+++ NEWS.md
@@ -1,3 +1,5 @@
+# guideR (development version)
+
 # guideR 0.10.0
 
 **New features**