+++ NEWS.md
@@ -2,8 +2,10 @@
 
 **New features**
 
-* new experimental functions `tbl_maihda()`, `tbl_strata_info()` and 
-  `tbl_strata_predictions()()` to display the results of a MAIHDA analysis
+* new experimental functions `tbl_maihda()`, `tbl_partially_adjusted_maihda()`,
+  `tbl_strata_info()`, `tbl_strata_predictions()()` and 
+  `plot_maihda_predictions_by()` to display  and plot the results of a MAIHDA
+  analysis (#66)
 * new utility `add_glance_header()` to add a group header before glance
   statistics added to a `gtsummary::tbl_regression()` table with 
   `gtsummary::add_glance_table()`